phase8/streamlit/requirements_windows.txt
streamlit==1.28.0
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
